requirements.txt
django~=5.1.6
Pillow~=11.1.0
djangorestframework-simplejwt~=5.4.0
drf-yasg~=1.21.9
drf-spectacular~=0.28.0